allow non-integer framerates for q3 shader animmap commands
git-svn-id: svn://svn.icculus.org/twilight/trunk/darkplaces@6850 d7cf8633-e32d-0410-b094-e92efae38249

+++ model_brush.c
@@ -4146,7 +4146,7 @@ static void Mod_Q3BSP_LoadShaders(void)
 						{
 							int i;
 							layer->numframes = min(numparameters - 2, TEXTURE_MAXFRAMES);
-							layer->framerate = atoi(parameter[1]);
+							layer->framerate = atof(parameter[1]);
 							for (i = 0;i < layer->numframes;i++)
 								strlcpy(layer->texturename[i], parameter[i + 2], sizeof(layer->texturename));
 						}
